Revert "wip: scaffold mint to instruction creator and test"
This reverts commit f8254050f8dd0c1b3d103277697be3d1037c8a5e.
