+++ associated-token-account/program/Cargo.toml
@@ -17,7 +17,6 @@ solana-program = "1.7.11"
 spl-token = { version = "3.2", path = "../../token/program", features = ["no-entrypoint"] }
 
 [dev-dependencies]
-solana-program = "1.7.11"
 solana-program-test = "1.7.11"
 solana-sdk = "1.7.11"
