--- associated-token-account/program/tests/program_test/cookies.rs
@@ -1,18 +1,0 @@
-use solana_program::pubkey::Pubkey;
-use solana_sdk::signature::Keypair;
-
-#[derive(Debug)]
-pub struct MintCookie {
-    pub address: Pubkey,
-    pub mint_authority: Keypair,
-}
-
-#[derive(Debug)]
-pub struct TokenAccountCookie {
-    pub address: Pubkey,
-}
-
-#[derive(Debug)]
-pub struct WalletCookie {
-    pub address: Pubkey,
-}
